pyproject.toml
[project]
name = "miloworks"
version = "0.1.0"
description = "Make AI shorts from chained-prompt manifests. Local studio for prompt-driven cartoon series."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "MiloWorks Studios" }]
keywords = ["ai", "video", "animation", "grok", "kling", "wan", "ffmpeg"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Multimedia :: Video",
    "Topic :: Artistic Software",
]
dependencies = [
    "xai-sdk>=1.11.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "rich>=13.9.0",
    "fal-client>=0.14.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
]

[project.urls]
Homepage = "https://miloworks.tv"
Repository = "https://github.com/MiloworksTV/MiloworksTV"
Issues = "https://github.com/MiloworksTV/MiloworksTV/issues"

[project.scripts]
miloworks = "miloworks.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["miloworks"]

[tool.hatch.build.targets.wheel.force-include]
"miloworks/web/index.html" = "miloworks/web/index.html"
"miloworks/workflows" = "miloworks/workflows"
